Move sync into sub-command
