+++ cmd/error.go
@@ -1,8 +1,6 @@
 package cmd
 
-import (
-	"github.com/giantswarm/microerror"
-)
+import "github.com/giantswarm/microerror"
 
 var invalidConfigError = &microerror.Error{
 	Kind: "invalidConfigError",
@@ -13,20 +11,11 @@ func IsInvalidConfig(err error) bool {
 	return microerror.Cause(err) == invalidConfigError
 }
 
-var invalidFlagError = &microerror.Error{
+var invalidFlagsError = &microerror.Error{
 	Kind: "invalidFlagsError",
 }
 
 // IsInvalidFlags asserts invalidFlagsError.
-func IsInvalidFlag(err error) bool {
-	return microerror.Cause(err) == invalidFlagError
-}
-
-var executionFailedError = &microerror.Error{
-	Kind: "executionFailedError",
-}
-
-// IsExecutionFailed asserts executionFailedError.
-func IsExecutionFailed(err error) bool {
-	return microerror.Cause(err) == executionFailedError
+func IsInvalidFlags(err error) bool {
+	return microerror.Cause(err) == invalidFlagsError
 }
